+++ src/me/pdthx/Requests/PaymentRequestRequest.java
@@ -5,7 +5,7 @@ public class PaymentRequestRequest {
 	public String UserId = "";
 	public String DeviceId = "";
 	public String RecipientUri = "";
-	public Double Amount = (double) 0;
+	public double Amount = 0;
 	public String Comments = "";
 	public String SecurityPin = "";
 }
